@@ -14,5 +14,5 @@ export interface DefaultBootInput {
     custom_launcher_selector?: string;
 }
 
-export type BootInput = ((UserWithEmail | UserWithUid) & DefaultBootInput & Any) | DefaultBootInput
+export type BootInput = ((UserWithEmail | UserWithUid) & DefaultBootInput & Any) | DefaultBootInput;
 
